+++ public/js/components/webtoonpage/BookKeywordItem.js
@@ -1,0 +1,7 @@
+const BookKeywordItem = keyword => `
+  <li class="books__keyword__list__item">
+    <button type="button">${keyword}</button>
+  </li>
+`;
+
+export default BookKeywordItem;
