+++ public/js/pages/index.js
@@ -1,5 +1,6 @@
 export { default as Home } from './Home.js';
 export { default as Webtoon } from './Webtoon.js';
+export { default as Viewer } from './Viewer.js';
 export { default as MyPage } from './MyPage.js';
 export { default as Login } from './Login.js';
 export { default as Signup } from './Signup.js';
