+++ scss/05_pages/_viewer.scss
@@ -70,19 +70,30 @@
       @include flexbox;
       gap: 18px;
 
-      button {
-        @include removeBtnStyle;
-        @include flexCenter;
-        font-size: 12px;
-        color: #999;
+      &__link {
+        display: block;
 
-        i {
-          font-size: 22px;
-          width: 18px;
-        }
+        button {
+          @include removeBtnStyle;
+          @include flexCenter;
+          font-size: 12px;
+          color: #999;
 
-        &:nth-child(2) i {
-          transform: rotate(180deg);
+          .arrow-left {
+            i {
+              font-size: 22px;
+              width: 18px;
+            }
+          }
+
+          .arrow-right {
+            transform: rotate(180deg);
+
+            i {
+              font-size: 22px;
+              width: 18px;
+            }
+          }
         }
       }
     }
